add bag support
