--- lance/config/perception.json
@@ -1,150 +1,0 @@
-{
-    "/**":
-    {
-        "ros__parameters":
-        {
-            "map_frame_id": "map",
-            "odom_frame_id": "odom",
-            "base_frame_id": "base_link",
-            "scan_topic": "/multiscan/lidar_scan",
-            "imu_topic": "/multiscan/imu",
-            "tag_usage_mode": 0,
-            "metrics_pub_freq": 10,
-            "robot_crop_filter":
-            {
-                "min": [ -0.405, -0.6, 0.0 ],
-                "max": [ 0.735, 0.6, 0.959 ]
-            },
-            "trajectory_filter":
-            {
-                "sampling_window_s": 0.6,
-                "min_filter_window_s": 0.4,
-                "thresh":
-                {
-                    "avg_linear_error": 0.01,
-                    "avg_angular_error": 0.02,
-                    "max_linear_deviation": 0.01,
-                    "max_angular_deviation": 0.02
-                }
-            },
-            "dlo":
-            {
-                "use_timestamps_as_init": false,
-                "gravity_align": false,
-                "adaptive_params":
-                {
-                    "use": false,
-                    "lpf_coeff": 0.8
-                },
-                "keyframe":
-                {
-                    "thresh_D": 1.0,
-                    "thresh_R": 30.0,
-                    "submap":
-                    {
-                        "knn": 5,
-                        "kcv": 5,
-                        "kcc": 5
-                    }
-                },
-                "initial_pose":
-                {
-                    "use": true,
-                    "position": [ 0.0, 0.0, 0.0 ],
-                    "orientation": [ 1.0, 0.0, 0.0, 0.0 ]
-                },
-                "voxel_filter":
-                {
-                    "scan":
-                    {
-                        "use": true,
-                        "res": 0.15
-                    },
-                    "submap":
-                    {
-                        "use": false,
-                        "res": 0.15
-                    },
-                    "adaptive_leaf_size":
-                    {
-                        "range_coeff": 0.063,
-                        "stddev_coeff": 0.064,
-                        "offset": -0.29,
-                        "floor": 0.04,
-                        "ceil": 0.5,
-                        "precision": 0.01
-                    }
-                },
-                "immediate_filter":
-                {
-                    "use": false,
-                    "range": 1,
-                    "thresh_proportion": 0.3
-                },
-                "imu":
-                {
-                    "use": true,
-                    "use_orientation": true,
-                    "calib_time": 3
-                },
-                "gicp":
-                {
-                    "num_threads": 4,
-                    "min_num_points": 500,
-                    "s2s":
-                    {
-                        "k_correspondences": 10,
-                        "max_correspondence_distance": 0.4,
-                        "max_iterations": 32,
-                        "transformation_epsilon": 0.01,
-                        "euclidean_fitness_epsilon": 0.01,
-                        "ransac":
-                        {
-                            "iterations": 5,
-                            "outlier_rejection_thresh": 0.1
-                        }
-                    },
-                    "s2m":
-                    {
-                        "k_correspondences": 15,
-                        "max_correspondence_distance": 0.5,
-                        "max_iterations": 32,
-                        "transformation_epsilon": 0.0009,
-                        "euclidean_fitness_epsilon": 0.005,
-                        "ransac":
-                        {
-                            "iterations": 6,
-                            "outlier_rejection_thresh": 0.05
-                        }
-                    }
-                }
-            },
-            "fiducial_detection":
-            {
-                "max_range": 1.0,
-                "plane_seg_thickness": 0.01,
-                "ground_seg_thickness": 0.02,
-                "up_vec_max_angular_dev": 1.0,
-                "planes_max_angular_dev": 0.2,
-                "vox_resolution": 0.03,
-                "max_proportion_leftover": 0.05,
-                "min_num_input_points": 300,
-                "min_plane_seg_points": 80
-            },
-            "mapping":
-            {
-                "frustum_search_radius": 0.04,
-                "radial_distance_thresh": 0.05,
-                "delete_delta_coeff": 0.08,
-                "delete_max_range": 3.5,
-                "add_max_range": 3.5,
-                "voxel_size": 0.1
-            },
-            "traversibility":
-            {
-                "chunk_horizontal_range": 3.5,
-                "chunk_vertical_range": 1.0
-            }
-        }
-    }
-}
