+++ launch-utils/src/launch_utils/actions.py
@@ -1,5 +1,4 @@
 import os
-import re
 import sys
 import json
 import yaml
@@ -64,12 +63,78 @@ def get_robot_state_pub_action(config):
         parameters = [config]
     )
 
-def get_bag_play_action():
-    return None
+# ----
 
-def get_bag_record_action():
-    return None
+def get_bag_play_action(
+        bag : str,
+        topics : list = [],
+        paused : bool = True,
+        loop : bool = False,
+        remappings : dict = {} ):
+    cmd_args = ['ros2', 'bag', 'play', '--clock', '10', bag]
+    if topics:
+        cmd_args.append('--topics')
+        cmd_args.extend(topics)
+    if paused:
+        cmd_args.append('--start-paused')
+    if loop:
+        cmd_args.append('--loop')
+    if remappings:
+        cmd_args.append('--remap')
+        for in_, out_ in remappings.items():
+            cmd_args.append(f'{in_}:={out_}')
 
+    return ExecuteProcess(
+        cmd = cmd_args,
+        output = 'screen'
+    )
+def get_bag_play_action_from_config(bag, config):
+    return get_bag_play_action(
+        bag,
+        config.get('topics', []),
+        config.get('start_paused', True),
+        config.get('loop', False),
+        config.get('remappings', {}))
+def add_bag_record_action(
+        topics : list,
+        file_prefix = 'bag_recordings/bag',
+        mcap = True ):
+    cmd_args = [
+        'ros2', 'bag', 'record',
+        '-o', f'{file_prefix}_{ datetime.now().strftime("%Y_%m_%d-%H_%M_%S") }' ]
+    if mcap:
+        cmd_args.append('-s')
+        cmd_args.append('mcap')
+    if topics:
+        cmd_args.extend(topics)
+    else:
+        cmd_args.append('--all')
+    ExecuteProcess(
+        cmd = cmd_args,
+        output = 'screen'
+    )
+def add_bag_rerecord_action(
+        src_bag : str,
+        exclude_topics : list = [],
+        mcap = True,
+        bag_name = ''):
+    cmd_args = [
+        'ros2', 'bag', 'record', '--all', '--use-sim-time',
+        '-o', (bag_name if bag_name else
+                f'{src_bag.rstrip("/")}-rerecord_{ datetime.now().strftime("%Y_%m_%d-%H_%M_%S") }') ]
+    if mcap:
+        cmd_args.append('-s')
+        cmd_args.append('mcap')
+    if exclude_topics:
+        cmd_args.append('--exclude')
+        cmd_args.append('|'.join(exclude_topics))
+    ExecuteProcess(
+        cmd = cmd_args,
+        output = 'screen'
+    )
+
+
+# ---
 
 def extract_util_configs(config):
     v = {}
@@ -78,7 +143,7 @@ def extract_util_configs(config):
             v[tag] = config[tag]
             del config[tag]
 
-def get_util_actions(config):
+def get_util_actions(config, launch_args = {}):
     a = []
     if 'robot_tf' in config:
         a.append(
@@ -98,92 +163,11 @@ def get_util_actions(config):
                 common.flatten_dict(config['joy_node'])
             )
         )
+    if 'bag' in launch_args:
+        a.append(
+            get_bag_play_action_from_config(
+                launch_args['bag'],
+                config.get('bag_play', {})
+            )
+        )
     return a
-
-
-# ----
-
-def get_bag_topic_types(bag : str):
-    topic_pattern = re.compile(r'Topic:\s+(\S+)\s+\|\s+Type:\s+(\S+)')  # "Parse lines like: '/topic_name [msg_type]'"
-    output = os.popen(f'ros2 bag info {bag}').read().rstrip()
-
-    type_topics = {}
-    for line in output.splitlines():
-        match = topic_pattern.search(line)
-        if match:
-            topic, msg_type = match.groups()
-            if msg_type not in type_topics:
-                type_topics[msg_type] = []
-            type_topics[msg_type].append(topic)
-
-    return type_topics
-
-def add_bag_play_action(
-        bag : str,
-        topics : list,
-        actions : list,
-        paused : bool = True,
-        loop : bool = False,
-        remappings : dict = {} ):
-    cmd_args = ['ros2', 'bag', 'play', '--clock', '10', bag]
-    if topics:
-        cmd_args.append('--topics')
-        cmd_args.extend(topics)
-    if paused:
-        cmd_args.append('--start-paused')
-    if loop:
-        cmd_args.append('--loop')
-    if remappings:
-        cmd_args.append('--remap')
-        for in_, out_ in remappings.items():
-            cmd_args.append(f'{in_}:={out_}')
-
-    actions.append(
-        ExecuteProcess(
-            cmd = cmd_args,
-            output = 'screen'
-        )
-    )
-def add_bag_record_action(
-        topics : list,
-        actions : list,
-        file_prefix = 'bag_recordings/bag',
-        mcap = True ):
-    cmd_args = [
-        'ros2', 'bag', 'record',
-        '-o', f'{file_prefix}_{ datetime.now().strftime("%Y_%m_%d-%H_%M_%S") }' ]
-    if mcap:
-        cmd_args.append('-s')
-        cmd_args.append('mcap')
-    if topics:
-        cmd_args.extend(topics)
-    else:
-        cmd_args.append('--all')
-    actions.append(
-        ExecuteProcess(
-            cmd = cmd_args,
-            output = 'screen'
-        )
-    )
-def add_bag_rerecord_action(
-        actions : list,
-        src_bag : str,
-        exclude_topics : list = [],
-        mcap = True,
-        bag_name = ''):
-    cmd_args = [
-        'ros2', 'bag', 'record', '--all', '--use-sim-time',
-        '-o', (bag_name if bag_name else
-                f'{src_bag.rstrip("/")}-rerecord_{ datetime.now().strftime("%Y_%m_%d-%H_%M_%S") }') ]
-    if mcap:
-        cmd_args.append('-s')
-        cmd_args.append('mcap')
-    if exclude_topics:
-        cmd_args.append('--exclude')
-        cmd_args.append('|'.join(exclude_topics))
-    actions.append(
-        ExecuteProcess(
-            cmd = cmd_args,
-            output = 'screen'
-        )
-    )
